Local Database : update method remove and replace

+++ GameDex/Helpers/Protocols/Database.swift
@@ -10,7 +10,7 @@ import Foundation
 // sourcery: AutoMockable
 protocol Database {
     func add(newEntity: SavedGame, platform: Platform, callback: @escaping (DatabaseError?) -> ())
-    func getPlatform(platformId: Int16) -> Result<PlatformCollected?, DatabaseError>
+    func getPlatform(platformId: Int) -> Result<PlatformCollected?, DatabaseError>
     func fetchAllPlatforms() -> Result<[PlatformCollected], DatabaseError>
     func replace(savedGame: SavedGame, callback: @escaping (DatabaseError?) -> ())
     func remove(savedGame: SavedGame, callback: @escaping (DatabaseError?) -> ())
